# cmd/mcp-server/docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "REST access to the Safecast radiation monitoring dataset via simplemap.safecast.org — 200M+ measurements from citizen scientists worldwide. All data is CC0-licensed and read-only. Powered by PostgreSQL+PostGIS.",
        "title": "Safecast Map API",
        "contact": {
            "name": "Safecast",
            "url": "https://safecast.org"
        },
        "license": {
            "name": "CC0 1.0 Universal",
            "url": "https://creativecommons.org/publicdomain/zero/1.0/"
        },
        "version": "1.0"
    },
    "host": "simplemap.safecast.org",
    "basePath": "/api",
    "paths": {
...
        "/spectrum/{marker_id}": {
            "get": {
                "description": "Returns complete gamma spectroscopy data including all channel counts for the specified measurement point. Use GET /api/spectra to find marker IDs.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "spectroscopy"
                ],
                "summary": "Get full spectroscopy channel data",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Marker/measurement identifier",
                        "name": "marker_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Full spectroscopy channel data",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Invalid marker_id",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "503": {
                        "description": "Database unavailable",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}